SectionManager: Rename to ConfigurationGathering
