+++ coalib/settings/ConfigurationGathering.py
